refactor: refactor tablecells to use named export
