--- src/components/ui/TableCells/CellBase/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./CellBase";
-export * from "./CellBase";
