--- src/components/ui/TableCells/ConnectionTypeCell/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ConnectionTypeCell";
-export * from "./ConnectionTypeCell";
