--- src/components/ui/TableCells/ModeCell/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ModeCell";
-export * from "./ModeCell";
