--- src/components/ui/TableCells/ModelDefinitionCell/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ModelDefinitionCell";
-export * from "./ModelDefinitionCell";
