--- src/components/ui/TableCells/ModelsCell/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ModelsCell";
-export * from "./ModelsCell";
