--- src/components/ui/TableCells/NameCell/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./NameCell";
-export * from "./NameCell";
