+++ src/components/ui/TableCells/index.ts
@@ -1,29 +1,12 @@
-import InstanceCell from "./InstanceCell";
-import type { InstanceCellProps } from "./InstanceCell";
-import NameCell from "./NameCell";
-import type { NameCellProps } from "./NameCell";
-import ConnectionTypeCell from "./ConnectionTypeCell";
-import type { ConnectionTypeCellProps } from "./ConnectionTypeCell";
-import ModeCell from "./ModeCell";
-import type { ModeCellProps } from "./ModeCell";
-import ModelsCell from "./ModelsCell";
-import type { ModelsCellProps } from "./ModelsCell";
-import ModelDefinitionCell from "./ModelDefinitionCell";
-import type { ModelDefintionCellProps } from "./ModelDefinitionCell";
-
-export {
-  InstanceCell,
-  NameCell,
-  ConnectionTypeCell,
-  ModeCell,
-  ModelsCell,
-  ModelDefinitionCell,
-};
-export type {
-  InstanceCellProps,
-  NameCellProps,
-  ConnectionTypeCellProps,
-  ModeCellProps,
-  ModelsCellProps,
-  ModelDefintionCellProps,
-};
+export { InstanceCell } from "./InstanceCell";
+export type { InstanceCellProps } from "./InstanceCell";
+export { NameCell } from "./NameCell";
+export type { NameCellProps } from "./NameCell";
+export { ConnectionTypeCell } from "./ConnectionTypeCell";
+export type { ConnectionTypeCellProps } from "./ConnectionTypeCell";
+export { ModeCell } from "./ModeCell";
+export type { ModeCellProps } from "./ModeCell";
+export { ModelsCell } from "./ModelsCell";
+export type { ModelsCellProps } from "./ModelsCell";
+export { ModelDefinitionCell } from "./ModelDefinitionCell";
+export type { ModelDefinitionCellProps } from "./ModelDefinitionCell";
